+++ CHANGELOG.md
@@ -3,6 +3,9 @@
 ## 2.0.0
 
 - [BREAKING] Returning a tuple and creating a ref object automatically
+- Using resize-observer-polyfill instead of resize-observer
+- Fixed an issue where resize observer would trigger changes endlessly
+- Added tests using Karma
 
 ## 1.0.0